bug fix, web3 and mpc asset rename

+++ src/main/java/com/cobo/custody/api/client/domain/asset/MPCAssetCoinInfo.java
@@ -2,6 +2,8 @@ package com.cobo.custody.api.client.domain.asset;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class MPCAssetCoinInfo {
     private String coin;
     @JsonProperty(value = "chain_code")
@@ -9,7 +11,7 @@ public class MPCAssetCoinInfo {
     @JsonProperty(value = "display_code")
     private String displayCode;
     private String description;
-    private Double amount;
+    private BigDecimal amount;
 
     public String getCoin() {
         return coin;
@@ -43,11 +45,11 @@ public class MPCAssetCoinInfo {
         this.description = description;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
